fixing_mifare_classic 0010

+++ smartCard/AcsReader.cpp
@@ -370,8 +370,8 @@ ApduResponse AcsReader::parseResponse(char *response, ulong responseLength, bool
         /* apduResponse.statusWord = (response[responseLength - 2] << 8) | response[responseLength - 1]; */
 
         // Extracción correcta de los últimos dos bytes
-        uint8_t highByte = static_cast<uint8_t>(response[responseLength - 2]);
-        uint8_t lowByte = static_cast<uint8_t>(response[responseLength - 1]);
+        ulong highByte = static_cast<ulong>(response[responseLength - 2]);
+        ulong lowByte = static_cast<ulong>(response[responseLength - 1]);
         apduResponse.statusWord = (highByte << 8) | lowByte;
 
         qDebug() << "Extracted statusWord (hex):" << QString::number(apduResponse.statusWord, 16).toUpper();
